requirements.txt
anyio==4.7.0
argcomplete==3.5.2
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
asgiref==3.8.1
asttokens==3.0.0
async-lru==2.0.4
attrs==24.3.0
babel==2.16.0
beautifulsoup4==4.12.3
bleach==6.2.0
blinker==1.9.0
build==1.2.2.post1
CacheControl==0.14.1
cattrs==24.1.2
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.0
cleo==2.1.0
click==8.1.8
comm==0.2.2
contourpy==1.3.1
crashtest==0.4.1
cryptography==44.0.0
cycler==0.12.1
debugpy==1.8.11
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.9
Django==5.1.3
django-environ==0.12.0
django-stubs==5.1.1
django-stubs-ext==5.1.1
djangorestframework==3.15.2
dulwich==0.21.7
executing==2.1.0
fastjsonschema==2.21.1
filelock==3.16.1
Flask==3.1.0
fonttools==4.55.3
fqdn==1.5.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
installer==0.7.0
ipykernel==6.29.5
ipython==8.31.0
ipywidgets==8.1.5
isoduration==20.11.0
itsdangerous==2.2.0
jaraco.classes==3.4.0
jedi==0.19.2
jeepney==0.8.0
Jinja2==3.1.5
json5==0.10.0
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter==1.1.1
jupyter-console==6.6.3
jupyter-events==0.11.0
jupyter-lsp==2.2.5
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_terminals==0.5.3
jupyterlab==4.3.4
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
jupyterlab_widgets==3.0.13
keyring==24.3.1
kiwisolver==1.4.7
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.10.0
matplotlib-inline==0.1.7
mistune==3.0.2
more-itertools==10.5.0
msgpack==1.1.0
nbclient==0.10.2
nbconvert==7.16.4
nbformat==5.10.4
nest-asyncio==1.6.0
nodeenv==1.9.1
notebook==7.3.2
notebook_shim==0.2.4
numpy==2.2.1
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
parso==0.8.4
pexpect==4.9.0
pillow==11.0.0
pipx==1.7.1
pkginfo==1.12.0
platformdirs==4.3.6
poetry==1.8.5
poetry-core==1.9.1
poetry-plugin-export==1.8.0
prometheus_client==0.21.1
prompt_toolkit==3.0.48
psutil==6.1.1
ptyprocess==0.7.0
pure_eval==0.2.3
pycparser==2.22
Pygments==2.18.0
pyparsing==3.2.0
pyproject_hooks==1.2.0
pyright==1.1.389
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-json-logger==3.2.1
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
RapidFuzz==3.11.0
referencing==0.35.1
requests==2.32.3
requests-cache==1.2.1
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.22.3
seaborn==0.13.2
SecretStorage==3.3.3
Send2Trash==1.8.3
setuptools==75.6.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
sqlparse==0.5.2
stack-data==0.6.3
terminado==0.18.1
tinycss2==1.4.0
tomlkit==0.13.2
tornado==6.4.2
traitlets==5.14.3
trove-classifiers==2024.10.21.16
types-python-dateutil==2.9.0.20241206
types-PyYAML==6.0.12.20240917
typing_extensions==4.12.2
tzdata==2024.2
uri-template==1.3.0
url-normalize==1.4.3
urllib3==2.3.0
userpath==1.9.2
virtualenv==20.28.0
wcwidth==0.2.13
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
widgetsnbextension==4.0.13
